[Feat/#61] - 커뮤니티 API 환경 세팅 & 상세 게시글 가져오기 연동

+++ data/src/main/java/com/umc/data/dataSource/remote/schedule/ScheduleRemoteDataSource.kt
@@ -8,7 +8,7 @@ import com.umc.domain.model.base.ApiState
 interface ScheduleRemoteDataSource {
     suspend fun getScheduleList(): ApiState<List<ScheduleListResponse>>
 
-    suspend fun getScheduleDetail(scheduleId: Int): ApiState<ScheduleDetailResponse>
+    suspend fun getScheduleDetail(scheduleId: Long): ApiState<ScheduleDetailResponse>
 
     suspend fun getMonthSchedule(year: Int, month: Int): ApiState<List<ScheduleMonthResponse>>
